+++ lld/MachO/SymbolTable.h
@@ -52,7 +52,7 @@ public:
   Symbol *addLazy(StringRef name, ArchiveFile *file,
                   const llvm::object::Archive::Symbol &sym);
 
-  Defined *addSynthetic(StringRef name, InputSection *, uint32_t value,
+  Defined *addSynthetic(StringRef name, InputSection *, uint64_t value,
                         bool isPrivateExtern, bool includeInSymtab);
 
   ArrayRef<Symbol *> getSymbols() const { return symVector; }
